refactor: rename Filters component to NavBar, other small improvements in NavBar
