+++ 1-ECommerce/src/App.tsx
@@ -1,10 +1,10 @@
 import "./index.css";
-import { Filters } from "./components/Filters"
+import { NavBar } from "./components/NavBar"
 
 export default function App() {
     return (
         <div className="contaiener py-8 w-full h-full lg:flex lg:flex-col">
-            <Filters />
+            <NavBar />
         </div>
     )
 }
